+++ src/helpers.ts
@@ -1,0 +1,28 @@
+export const filterParams = (k: string) => k !== '__params' && k !== '__alias'
+
+export const getParams = (params: any) => {
+  if (!params) {
+    return ''
+  }
+  const variables = Object.keys(params)
+    .map(key => `${key}: ${params[key]}`)
+    .join(', ')
+  return `(${variables})`
+}
+
+// TODO: Tail Call Recursion
+export const joinFieldRecursively = (fieldOrObject: any): string => {
+  const joinedFields = Object.keys(fieldOrObject)
+    .filter(filterParams)
+    .map(key => {
+      if (Array.isArray(fieldOrObject)) {
+        return `${joinFieldRecursively(fieldOrObject[0])}`
+      }
+      if (typeof fieldOrObject[key] === 'object') {
+        return `${key} { ${joinFieldRecursively(fieldOrObject[key])} }`
+      }
+      return key
+    })
+    .join(' ')
+  return joinedFields
+  }
